% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralPatterns.R
\name{spectralPatterns}
\alias{spectralPatterns}
\title{Pattern Detection of Single Spectra on their Multi-components}
\usage{
spectralPatterns(spectra)
}
\arguments{
\item{spectra}{takes in a dataframe of spectra, this version accepts three-bases only. They should be organized as follows: C1, C2, C3, C1+C2, C1+C3, C2+C3, C1+C2+C3}
}
\description{
This function will take in a set of spectra (single components, and their respective multi-components) and return patterns detected based on correlation results of three extracted features (means, variance, and turn-points).
}
\examples{
cat_function()
}
\keyword{correlation}
\keyword{detection,}
\keyword{pattern}
\keyword{spectra,}

